Don't Install CoC Extensions From Source
This is creating problems on Linux due to sudo priveleges. Because the
frozen lockfile is used, the plugins fail the installation but appear to
be 'installed' when `:PlugInstall` is called a second time. Simply move
the set of desired extensions to `g:coc_global_extensions` and CoC will
manage installation and updates on starting Neovim.